requirements.txt
Django==4.2.19
PyMySQL==1.1.0
Pillow>=10.0.0
redis==7.1.0
whitenoise==6.11.0

